% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster the recurrence embedding via HDBSCAN}
\usage{
cluster(object, min_points = 10)
}
\arguments{
\item{object}{The recurrence embedding object returned by \code{\link[=embed]{embed()}}.}

\item{min_points}{The \code{min_points} argument of \code{\link[dbscan:hdbscan]{dbscan::hdbscan()}}.}
}
\description{
This function extends the embedding by an HDBSCAN clustering. Note that the
\code{embedding} item is extended by two columns \code{cluster} and \code{membership_prob}
of the as optimal identified cut in the hierarchical clustering. The full
HDBSCAN object returned by \code{\link[dbscan:hdbscan]{dbscan::hdbscan()}} is returned as well.
}
\examples{
times <- recur::synthetic
measured <- measure(data = times, size = "triangle", shape = "wide")
embedded <- embed(object = measured)
embedded_and_clustered <- cluster(embedded, min_points = 25)

plot(embedded_and_clustered$embedding$x, 
     embedded_and_clustered$embedding$y,
     col = 1 + embedded_and_clustered$embedding$cluster)
     
plot(embedded_and_clustered$hdbscan)

}
\references{
Leland McInnes, John Healy, Steve Astels. \href{https://hdbscan.readthedocs.io}{The hdbscan Clustering Library}.

Ricardo J.G.B. Campello, Davoud Moulavi, Joerg Sander (2013).
\href{https://link.springer.com/chapter/10.1007/978-3-642-37456-2_14}{Density-Based Clustering Based on Hierarchical Density Estimates},
PAKDD 2013, Part II, LNAI 7819, pp. 160-172.
}
\seealso{
\code{\link[=measure]{measure()}}, \code{\link[=embed]{embed()}}, \code{\link[dbscan:hdbscan]{dbscan::hdbscan()}}
}
