% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{Embed the recurrence table via UMAP}
\usage{
embed(object = NULL, data = NULL, ..., umap_object = TRUE)
}
\arguments{
\item{object}{The recurrence object returned by \code{\link[=measure]{measure()}}. If not provided,
\code{data} will be used to first compute the recurrence object before it is
embedded.}

\item{data}{A dataframe with columns "id", "date", "value". See \code{\link[=measure]{measure()}}
for more information. If provided while \code{object} is left empty, the
recurrence table is computed internally via \code{\link[=measure]{measure()}} first.}

\item{...}{Further parameters provided to \code{\link[=measure]{measure()}}.}

\item{umap_object}{If \code{TRUE} (default), the object returned by \code{\link[umap:umap]{umap::umap()}} is
returned as well.}
}
\value{

}
\description{
This function takes the recurrence table version of a set of time series
to embed the time series via \href{https://umap-learn.readthedocs.io/en/latest/#}{Uniform Manifold Approximation and Projection}
(UMAP).
This only provides access to the naive settings of the \link[umap:umap]{umap::umap} package.
}
\examples{
times <- recur::synthetic
measured <- measure(data = times, size = "triangle", shape = "wide")
embedded <- embed(object = measured)

# this generates the same result as
embedded <- embed(data = times, size = "triangle")
}
\references{
L. McInnes, J. Healy, J. Melville (2018).
\href{https://arxiv.org/abs/1802.03426}{\emph{UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction}},
ArXiv e-prints 1802.03426.
}
\seealso{
\code{\link[=measure]{measure()}}, \code{\link[umap:umap]{umap::umap()}}, \code{\link[=cluster]{cluster()}}
}
