% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Measure the recurrence within time series}
\usage{
measure(
  data,
  absolute = TRUE,
  size = c("triangle", "square"),
  shape = c("wide", "long"),
  start_date = NA,
  end_date = NA
)
}
\arguments{
\item{data}{A dataframe with columns "id", "date", "value". See \link{synthetic}
for more information.}

\item{absolute}{If TRUE (default), the difference between two observations is
the absolute difference.}

\item{size}{Should half of the recurrence observations be dropped from each
time series' matrix? If "triangle" (default), only the lower triangle
of the recurrence matrix of each time series is returned. If "square",
the full recurrence matrix is returned. This could for example be used
with \code{absolute = FALSE}.}

\item{shape}{If "wide" (default), the recurrence matrices are returned as a
wide dataframe which can be passed directly to, for example,
\code{\link[umap:umap]{umap::umap()}} for further modeling. If "long", the recurrence matrices
are returned as one long dataframe which makes use with \link[ggplot2:ggplot]{ggplot2::ggplot}
simple.}

\item{start_date}{The first date considered when extending every time series.
If \code{NA}, then set to \code{min(data$date)}.}

\item{end_date}{The last date considered when extending every time series.
If \code{NA}, then set to \code{max(data$date)}.}
}
\description{
Measure the recurrence within time series
}
\examples{
times <- recur::synthetic
measured <- measure(data = times, size = "triangle", shape = "wide")

}
\references{
Xixi Lia, Yanfei Kanga, Feng Li (2020).
\href{https://arxiv.org/abs/1904.08064}{\emph{Forecasting with Time Series Imaging}},
ArXiv e-prints 1802.03426.

J.-P. Eckmann, S. Oliffson Kamphorst, D. Ruelle (1987).
\href{http://iopscience.iop.org/0295-5075/4/9/004}{\emph{Recurrence Plots of Dynamical Systems}},
EPL (Europhysics Letters), Vol. 4, 9.
}
