% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic}
\alias{synthetic}
\title{900 monthly time series of length 48}
\format{A data frame with 43200 rows and 3 variables:
\describe{
\item{id}{the id of the time series in the format \code{ts-000}}
\item{date}{the date of the monthly observation, formatted as \code{YYYY-MM-01}}
\item{value}{the value of the time series (here i.i.d standard normal
random values)}
}}
\usage{
synthetic
}
\description{
A dataset containing 900 synthetic example time series that can be used
to trial the \code{recur} package functionality.
}
\keyword{datasets}
